% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_validation_response}
\alias{format_validation_response}
\title{Truncate the results of schematic validation}
\usage{
format_validation_response(x, n = 50)
}
\arguments{
\item{x}{Schematic validation result}

\item{n}{Number of records to keep}
}
\description{
Truncate the results of schematic validation
}
