% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{get_asset_view_table}
\alias{get_asset_view_table}
\title{/storage/asset/table}
\usage{
get_asset_view_table(
  url = "http://localhost:3001/v1/storage/assets/tables",
  access_token,
  asset_view,
  return_type = "json"
)
}
\arguments{
\item{url}{URL to schematic API endpoint}

\item{access_token}{synapse PAT}

\item{asset_view}{Synapse ID of asset view}
}
\description{
/storage/asset/table
}
