% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{get_synapse_table_names}
\alias{get_synapse_table_names}
\title{Get column names from a Synapse table
https://rest-docs.synapse.org/rest/GET/entity/id/table/query/async/get/asyncToken.html
Uses a table query to get the column names from a Synapse table}
\usage{
get_synapse_table_names(id, auth)
}
\arguments{
\item{id}{Synapse ID of table}

\item{auth}{Synapse token}
}
\description{
Get column names from a Synapse table
https://rest-docs.synapse.org/rest/GET/entity/id/table/query/async/get/asyncToken.html
Uses a table query to get the column names from a Synapse table
}
