% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{graph_by_edge_type}
\alias{graph_by_edge_type}
\title{graph by edge type endpoint}
\usage{
graph_by_edge_type(
  url = "https://schematic-dev.api.sagebionetworks.org/v1/schemas/get/graph_by_edge_type",
  schema_url,
  relationship = "requiresDependency",
  data_model_labels = "class_label"
)
}
\arguments{
\item{url}{URL of schematic API endpoint}

\item{schema_url}{URL of data model}

\item{relationship}{Argument to schematic graph_by_edge_type}

\item{data_model_labels}{Which labels to use (class_label, display_label)}
}
\description{
graph by edge type endpoint
}
