% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{manifest_download}
\alias{manifest_download}
\title{Download a manifest}
\usage{
manifest_download(
  url = "http://localhost:3001/v1/manifest/download",
  access_token,
  manifest_id,
  as_json = TRUE,
  new_manifest_name = NULL
)
}
\arguments{
\item{url}{URI of API endpoint}

\item{access_token}{Synapse PAT}

\item{manifest_id}{the parent ID of the manifest}

\item{as_json}{if True return the manifest in JSON format}

\item{new_manifest_name}{Name of new manifest (Default = NULL)}

\item{asset_view}{ID of view listing all project data assets}
}
\value{
a csv of the manifest
}
\description{
Download an existing manifest
}
