% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{manifest_populate}
\alias{manifest_populate}
\title{Populate a manifest sheet}
\usage{
manifest_populate(
  url = "http://localhost:3001/v1/manifest/populate",
  schema_url,
  data_type,
  title,
  return_excel = FALSE,
  data_model_labels = "class_label",
  csv_file
)
}
\arguments{
\item{url}{URL to schematic API endpoint}

\item{schema_url}{URL to a schema jsonld}

\item{data_type}{Type of dataset}

\item{title}{Title of csv}

\item{csv_file}{Filepath of csv to validate}
}
\description{
Populate a manifest sheet
}
