% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{manifest_validate}
\alias{manifest_validate}
\title{schematic rest api to validate metadata}
\usage{
manifest_validate(
  url = "http://localhost:3001/v1/model/validate",
 
    schema_url = "https://raw.githubusercontent.com/ncihtan/data-models/main/HTAN.model.jsonld",
  data_type,
  csv_file
)
}
\arguments{
\item{url}{URL to schematic API endpoint}

\item{schema_url}{URL to a schema jsonld}

\item{data_type}{Type of dataset}

\item{csv_file}{Filepath of csv to validate}
}
\value{
An empty list() if sucessfully validated. Or a list of errors.
}
\description{
schematic rest api to validate metadata
}
