% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{model_component_requirements}
\alias{model_component_requirements}
\title{Given a source model component (see https://w3id.org/biolink/vocab/category for definnition of component), return all components required by it.}
\usage{
model_component_requirements(
  url = "http://localhost:3001/v1/model/component-requirements",
  schema_url,
  source_component,
  as_graph = FALSE
)
}
\arguments{
\item{schema_url}{Data Model URL}

\item{source_component}{an attribute label indicating the source component. (i.e. Patient, Biospecimen, ScRNA-seqLevel1, ScRNA-seqLevel2)}

\item{as_graph}{if False return component requirements as a list; if True return component requirements as a dependency graph (i.e. a DAG)}
}
\value{
A list of required components associated with the source component.
}
\description{
Given a source model component (see https://w3id.org/biolink/vocab/category for definnition of component), return all components required by it.
}
