% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_access}
\alias{synapse_access}
\title{Check Access Permissions to a Synapse Entity}
\usage{
synapse_access(
  url = "https://repo-prod.prod.sagebase.org/repo/v1/entity",
  id,
  access,
  auth
)
}
\arguments{
\item{url}{URL to REST API endpoint}

\item{id}{Synapse ID}

\item{access}{Access Type to check}

\item{auth}{Synapse authentication token}
}
\description{
wrapper for https://rest-docs.synapse.org/rest/GET/entity/id/access.html
}
