% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_download_file_handle}
\alias{synapse_download_file_handle}
\title{Download a synapse file from its URL
https://rest-docs.synapse.org/rest/GET/file/id.html}
\usage{
synapse_download_file_handle(dataFileHandleId, id, auth, filepath = NULL)
}
\arguments{
\item{dataFileHandleId}{The dataFileHandleId from an entity}

\item{id}{The synapse ID of the file to download}

\item{auth}{Synapse token}

\item{filepath}{Optional path to download data. If NULL, return a data frame.}
}
\description{
Download a synapse file from its URL
https://rest-docs.synapse.org/rest/GET/file/id.html
}
