% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_entity_children}
\alias{synapse_entity_children}
\title{Get children of a synapse entity
https://rest-docs.synapse.org/rest/POST/entity/children.html}
\usage{
synapse_entity_children(
  url = "https://repo-prod.prod.sagebase.org/repo/v1/entity/children",
  auth,
  parentId = NULL,
  nextPageToken = NULL,
  includeTypes = "project",
  sortBy = "NAME",
  sortDirection = "ASC",
  includeTotalChildCount = FALSE,
  includeSumFileSizes = FALSE
)
}
\arguments{
\item{url}{Synapse api endpoint}

\item{auth}{Synapse token}

\item{parentId}{Synapse ID of parent folder}

\item{nextPageToken}{Synapse next page token}

\item{includeTypes}{Types to return}

\item{sortBy}{Variable to sort by}

\item{sortDirection}{sort direction}

\item{includeTotalChildCount}{boolean include count of children}

\item{includeSumFileSizes}{boolean include sum of file sizes}
}
\description{
Get children of a synapse entity
https://rest-docs.synapse.org/rest/POST/entity/children.html
}
