% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_get}
\alias{synapse_get}
\title{GET Synapse Entity}
\usage{
synapse_get(
  url = "https://repo-prod.prod.sagebase.org/repo/v1/entity/",
  id,
  auth
)
}
\arguments{
\item{url}{URL of synapse REST API GET table entity endpoint}

\item{id}{ID of synapse table}

\item{auth}{Synapse PAT}
}
\description{
Wrapper for https://rest-docs.synapse.org/rest/GET/entity/id.html
}
