% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_get_project_scope}
\alias{synapse_get_project_scope}
\title{Get projects within scope of Synapse project}
\usage{
synapse_get_project_scope(
  url = "https://repo-prod.prod.sagebase.org/repo/v1/entity/",
  id,
  auth
)
}
\arguments{
\item{url}{Synapse api endpoint}

\item{id}{Synapse ID}

\item{auth}{Synapse token}
}
\description{
Get projects within scope of Synapse project
}
