% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_is_certified}
\alias{synapse_is_certified}
\title{Is a Synapse user certified?}
\usage{
synapse_is_certified(
  url = "https://repo-prod.prod.sagebase.org/repo/v1/user",
  endpoint = "certifiedUserPassingRecord",
  auth = NULL
)
}
\arguments{
\item{url}{URL to Synapse REST API for certification}

\item{endpoint}{The Synapse API endpoint}

\item{auth}{Synapse PAT or authorization token}
}
\description{
Is a Synapse user certified?
}
\details{
https://rest-docs.synapse.org/rest/GET/user/id/certifiedUserPassingRecord.html
Based on python client https://github.com/Sage-Bionetworks/synapsePythonClient/blob/3da21b9e2542c7da8cdab9925926737fe2162a54/synapseclient/client.py#L606
}
