% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_projects_user}
\alias{synapse_projects_user}
\title{Get projects a user has access to}
\usage{
synapse_projects_user(
  url = "https://repo-prod.prod.sagebase.org/repo/v1/projects/user",
  auth,
  nextPageToken = NULL
)
}
\arguments{
\item{url}{Synapse api endpoint}

\item{auth}{Synapse token}

\item{nextPageToken}{Synapse next page token}
}
\description{
Get projects a user has access to
}
