% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_storage_projects}
\alias{synapse_storage_projects}
\title{Get storage projects within a Synapse table
https://rest-docs.synapse.org/rest/GET/entity/id/table/query/async/get/asyncToken.html}
\usage{
synapse_storage_projects(
  id,
  auth,
  select_cols = c("id", "name", "parentId", "projectId", "type", "columnType")
)
}
\arguments{
\item{id}{Synapse ID of table}

\item{auth}{Synapse token}

\item{select_cols}{Columns to get from table}
}
\description{
Get storage projects within a Synapse table
https://rest-docs.synapse.org/rest/GET/entity/id/table/query/async/get/asyncToken.html
}
