% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_rest_api.R
\name{synapse_user_profile}
\alias{synapse_user_profile}
\title{Get Synapse user profile info}
\usage{
synapse_user_profile(
  url = "https://repo-prod.prod.sagebase.org/repo/v1/userProfile",
  auth = NULL
)
}
\arguments{
\item{url}{Synapse REST API url for userProfile}

\item{auth}{Synapse PAT or authorization token}
}
\description{
Get Synapse user profile info
}
\details{
For information on authenticating synapse REST API queries
https://docs.synapse.org/rest/#org.sagebionetworks.auth.controller.AuthenticationController
Authentication to Synapse services requires an access token passed in the
HTTP Authorization header, as per the HTTP bearer authorization standard.
https://datatracker.ietf.org/doc/html/rfc6750#section-2.1
https://docs.synapse.org/rest/#org.sagebionetworks.repo.web.controller.UserProfileController
PAT is the synapse personal access token OR the browser cookie from a logged-in session.
}
