% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{manifest_download}
\alias{manifest_download}
\title{Download an existing manifest}
\usage{
manifest_download(
  url = "http://localhost:3001/v1/manifest/download",
  input_token,
  asset_view,
  dataset_id,
  as_json = TRUE
)
}
\arguments{
\item{url}{URI of API endpoint}

\item{input_token}{Synapse PAT}

\item{asset_view}{ID of view listing all project data assets}

\item{dataset_id}{the parent ID of the manifest}

\item{as_json}{if True return the manifest in JSON format}
}
\value{
a csv of the manifest
}
\description{
Download an existing manifest
}
