% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{manifest_generate}
\alias{manifest_generate}
\title{schematic rest api to generate manifest}
\usage{
manifest_generate(
  url = "http://localhost:3001/v1/manifest/generate",
  schema_url =
    "https://raw.githubusercontent.com/ncihtan/data-models/main/HTAN.model.jsonld",
  title,
  data_type,
  oauth = "true",
  use_annotations = "false",
  dataset_id = NULL,
  asset_view,
  output_format
)
}
\arguments{
\item{title}{Name of dataset}

\item{data_type}{Type of dataset}

\item{oauth}{true or false STRING passed to python}

\item{use_annotations}{true or false STRING passed to python}

\item{dataset_id}{Synapse ID of existing manifest}
}
\value{
a URL to a google sheet
}
\description{
schematic rest api to generate manifest
}
