% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schematic_rest_api.R
\name{model_submit}
\alias{model_submit}
\title{schematic rest api to submit metadata}
\usage{
model_submit(
  url = "http://localhost:3001/v1/model/submit",
  schema_url,
  data_type,
  dataset_id,
  restrict_rules = FALSE,
  access_token,
  json_str = NULL,
  asset_view,
  manifest_record_type = "table_and_file",
  file_name,
  table_manipulation = "replace",
  hide_blanks = FALSE,
  table_column_names = "class_label",
  annotation_keys = "class_label",
  data_model_labels = "class_label",
  upload_file_annotation = TRUE
)
}
\arguments{
\item{url}{URL to schematic API endpoint}

\item{schema_url}{URL to a schema jsonld}

\item{data_type}{Type of dataset}

\item{dataset_id}{Synapse ID of existing manifest}

\item{restrict_rules}{Default = FALSE}

\item{access_token}{Synapse login cookie, PAT, or API key}

\item{json_str}{Json string to submit}

\item{asset_view}{Synapse fileview}

\item{manifest_record_type}{Default = "table_and_file"}

\item{file_name}{Name of file}

\item{table_manipulation}{Default = "replace"}

\item{hide_blanks}{Default = FALSE}

\item{table_column_names}{Default = "class_and_label"}

\item{annotation_keys}{Default = "class_and_label"}

\item{data_model_labels}{Default = "class_and_label"}

\item{upload_file_annotations}{Default = TRUE}
}
\value{
TRUE if successful upload or validate errors if not.
}
\description{
schematic rest api to submit metadata
}
